Add autocompletion for the gem command.
If using RVM, add rvm_leave_gem_alone=1 to ~/.rvmrc so gem is a file and
not a shell command.

This caches a remote list to ~/.remote-gem-cache.  It probably makes
sense to delete this file every now and then, either with cron or
launchd.

    <?xml version="1.0" encoding="UTF-8"?>
    <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
    <plist version="1.0">
    <dict>
      <key>Label</key>
      <string>org.joshuaclayton.clean_rubygems_cache</string>
      <key>ProgramArguments</key>
      <array>
        <string>/bin/rm</string>
        <string>/Users/joshuaclayton/.remote-gem-cache</string>
      </array>
      <key>StartCalendarInterval</key>
      <dict>
        <key>Hour</key>
        <integer>12</integer>
      </dict>
      <key>UserName</key>
      <string>joshuaclayton</string>
      <key>WorkingDirectory</key>
      <string>/Users/joshuaclayton</string>
    </dict>
    </plist>

This plist would clear out the file every day at noon.